Implement GetWebinarPanelist endpoint
